# api-v10/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pueblos magicos",
        "description": "Descripción de tu API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/registrar": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crea un nuevo usuario",
                "operationId": "475e0c7db715ee91a6d776e498efa7c4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "user_name": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "apellido_pat": {
                                                "type": "string"
                                            },
                                            "apellido_mat": {
                                                "type": "string"
                                            },
                                            "id_tipo_user": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "data": {
                                        "user_name": "lleon@ipn.mx",
                                        "password": "pruebas1",
                                        "nombre": "Luis",
                                        "apellido_pat": "Leon",
                                        "apellido_mat": "HDZ",
                                        "id_tipo_usuario": 2
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Inicio de session de un usuario",
                "operationId": "e7a377530f4b250da42ad2b352677c70",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "user_name": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "data": {
                                        "user_name": "bajimeneza@ipn.mx",
                                        "password": "Pruebas1"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cerrar sesión de un usuario",
                "operationId": "3f4630f2faa114f9a6495b04588632ec",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "Modelo para representar un usuario",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Calle 123"
                    },
                    "ciudad": {
                        "type": "string",
                        "example": "Ciudad de México"
                    },
                    "cp": {
                        "type": "integer",
                        "example": 12345
                    },
                    "id_tipo_user": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_estado": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "example": "juan.perez@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password_seguro"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-04-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-04-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}